(fix) Rename end_check to comply with style guide
